fix(cloud): reject duplicate analytics metadata flags
Validate all includeMetadata query values before analytics reads or export generation, and consolidate the route-boundary coverage into the existing analytics export suite.
